moved some components, work on sort
