+++ src/SpellBinder.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import LeftSide from "./components/SpellBinder/LeftSide.js";
-import RightSide from "./components/SpellBinder/RightSide.js";
+import RightSide from "./components/RightSide.js";
 import axios from "axios";
 import AppHeader from "./components/AppHeader.js";
 import InfoBar from "./components/InfoBar.js";
@@ -54,7 +54,7 @@ export function Spellbinder(props) {
                 <InfoBar examples={examples} instructions={instructions} />
                 <LeftSide className="LeftSide" setModels={setModels} models={models} points={points} setPoints={setPoints} wrongFirst={wrongFirst} wrongMiddle={wrongMiddle} wrongLast={wrongLast} setOver={setOver} question={question} counter={counter} setCounter={setCounter} submitted={submitted} setSubmitted={setSubmitted} value={value} setValue={setValue} score={score} setScore={setScore} correctWords={correctWords} setCorrectWords={setCorrectWords}></LeftSide>
             </div>
-            <RightSide score={score} correctWords={correctWords} className="RightSide"></RightSide>
+            <RightSide score={score} correctWords={correctWords} className="RightSide" />
         </div>
     );
 }
